src/requirements.txt
flask
uwsgi
passlib
python-dotenv
requests
sqlalchemy
marshmallow
marshmallow-sqlalchemy
flask-marshmallow
sqlacodegen
flask-wtf
flask-uploads
wtforms_alchemy
flask-session
flask-login
flask_sqlalchemy
flask_bootstrap
redis
pyyaml
pyodbc
Flask-RESTPlus
cryptography
googletrans